Tests/LibWeb/Text/expected/HTML/preserve-leading-bom-in-string-parsing.txt

DOMParser: length=4, codePoints=feff,54,68,65
innerHTML: length=4, codePoints=feff,54,68,65
